LayoutTests/inspector/console/console-dir-expected.txt

CONSOLE MESSAGE: line 9: test1,test2
CONSOLE MESSAGE: line 10: [object NodeList]
CONSOLE MESSAGE: line 11: [object XPathResult]
Tests that console logging dumps proper messages.

Array[2]
    0 = "test1"
    1 = "test2"
    length = 2
    __proto__ = Array[0]
NodeList[1]
    0 = HTMLHtmlElement
    constructor = NodeListConstructor
    length = 1
    __proto__ = NodeListPrototype
XPathResult
    booleanValue = Error: TYPE_ERR: DOM XPath Exception 52
    constructor = XPathResultConstructor
    invalidIteratorState = false
    numberValue = Error: TYPE_ERR: DOM XPath Exception 52
    resultType = 4
    singleNodeValue = Error: TYPE_ERR: DOM XPath Exception 52
    snapshotLength = Error: TYPE_ERR: DOM XPath Exception 52
    stringValue = Error: TYPE_ERR: DOM XPath Exception 52
    __proto__ = XPathResultPrototype
